dev: Refactor cache files
- Remove any circular import references.
- Move interfaces next to the supporting code.
