+++ packages/cspell/src/lint/index.ts
@@ -1,1 +1,2 @@
 export { runLint } from './lint';
+export { LintRequest } from './LintRequest';
