+++ packages/cspell/src/options.ts
@@ -1,4 +1,6 @@
-export interface LinterOptions extends BaseOptions {
+import type { CacheOptions } from './util/cache';
+
+export interface LinterOptions extends BaseOptions, CacheOptions {
     /**
      * Display verbose information
      */
@@ -39,24 +41,6 @@ export interface LinterOptions extends BaseOptions {
      * Show suggestions for spelling errors.
      */
     showSuggestions?: boolean;
-
-    /**
-     * Store the info about processed files in order to only operate on the changed ones.
-     */
-    cache?: boolean;
-
-    // cspell:word cspellcache
-    /**
-     * Path to the cache location. Can be a file or a directory.
-     * If none specified .cspellcache will be used.
-     * The file will be created in the directory where the cspell command is executed.
-     */
-    cacheLocation?: string;
-
-    /**
-     * Strategy to use for detecting changed files, default: metadata
-     */
-    cacheStrategy?: 'metadata' | 'content';
 
     /**
      * Enable filtering out files matching globs found in `.gitignore` files.
